+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/dialog/ComposableSelectDialog.kt
@@ -2,19 +2,17 @@ package com.flipperdevices.filemanager.impl.composable.dialog
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Divider
 import androidx.compose.material.Text
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
+import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 
@@ -50,11 +48,7 @@ private fun ComposableDialogOption(
     onClick: () -> Unit
 ) {
     Text(
-        modifier = Modifier.clickable(
-            interactionSource = MutableInteractionSource(),
-            indication = rememberRipple(),
-            onClick = onClick
-        ).then(modifier),
+        modifier = Modifier.clickableRipple(onClick = onClick).then(modifier),
         text = stringResource(textId),
         style = LocalTypography.current.titleM18
     )