+++ components/keyscreen/shared/src/main/java/com/flipperdevices/keyscreen/shared/bar/ComposableBarSimpleText.kt
@@ -1,13 +1,10 @@
 package com.flipperdevices.keyscreen.shared.bar
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.material.Text
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 
@@ -21,11 +18,7 @@ fun ComposableBarSimpleText(
     var textModifier = modifier
 
     if (onClick != null) {
-        textModifier = textModifier.clickable(
-            interactionSource = remember { MutableInteractionSource() },
-            indication = rememberRipple(bounded = false),
-            onClick = onClick
-        )
+        textModifier = textModifier.clickableRipple(bounded = false, onClick = onClick)
     }
 
     Text(