+++ components/nfc/attack/impl/src/main/java/com/flipperdevices/nfc/attack/impl/composable/elements/MifareClassicComposable.kt
@@ -2,8 +2,6 @@ package com.flipperdevices.nfc.attack.impl.composable.elements
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -15,7 +13,6 @@ import androidx.compose.material.Card
 import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -25,6 +22,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
@@ -45,11 +43,7 @@ fun MifareClassicComposable(
             val hasMfKey32Notification by nfcAttackViewModel.hasMfKey32Notification()
                 .collectAsState()
             MifareClassicMfKey32(
-                modifier = Modifier.clickable(
-                    interactionSource = MutableInteractionSource(),
-                    indication = rememberRipple(),
-                    onClick = onOpenMfKey32
-                ),
+                modifier = Modifier.clickableRipple(onClick = onOpenMfKey32),
                 hasNotification = hasMfKey32Notification
             )
         }