+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/composable/category/VersionCategory.kt
@@ -1,7 +1,6 @@
 package com.flipperdevices.settings.impl.composable.category
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -28,16 +27,12 @@ fun VersionCategory(version: String, onActivateExpertMode: () -> Unit) {
     var howMuchClick by remember { mutableStateOf(0) }
     val versionText = "${stringResource(id = R.string.version)}: $version"
     Column(
-        modifier = Modifier.fillMaxWidth().clickable(
-            interactionSource = MutableInteractionSource(),
-            indication = null,
-            onClick = {
-                howMuchClick++
-                if (howMuchClick > EXPERT_MODE_CLICK_COUNT) {
-                    onActivateExpertMode()
-                }
+        modifier = Modifier.fillMaxWidth().clickable {
+            howMuchClick++
+            if (howMuchClick > EXPERT_MODE_CLICK_COUNT) {
+                onActivateExpertMode()
             }
-        ),
+        },
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {