+++ components/share/uploader/src/main/kotlin/com/flipperdevices/uploader/compose/content/ComposableSheetPending.kt
@@ -5,7 +5,6 @@ import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -69,11 +68,7 @@ private fun ComposableSheetAction(
                 .size(48.dp)
                 .clip(CircleShape)
                 .background(LocalPallet.current.shareSheetBackgroundAction.copy(alpha = 0.1f))
-                .clickable(
-                    interactionSource = MutableInteractionSource(),
-                    indication = null,
-                    onClick = onAction
-                ),
+                .clickable(onClick = onAction),
             contentAlignment = Alignment.Center
         ) {
             Image(