+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableFirmwareUpdaterError.kt
@@ -3,22 +3,19 @@ package com.flipperdevices.updater.card.composable
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
-import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.flipperdevices.core.ui.ktx.clickableRipple
 import com.flipperdevices.core.ui.ktx.painterResourceByKey
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
@@ -61,11 +58,7 @@ fun ComposableFirmwareUpdaterError(
 
     Text(
         modifier = Modifier
-            .clickable(
-                indication = rememberRipple(),
-                onClick = retryUpdate,
-                interactionSource = remember { MutableInteractionSource() }
-            )
+            .clickableRipple(onClick = retryUpdate)
             .fillMaxWidth()
             .padding(top = 4.dp, bottom = 8.dp),
         text = stringResource(R.string.updater_card_updater_error_retry),